chore(auth): 优化双Token设计，版本号+黑名单机制

+++ fitness-system/src/main/java/com/wsw/fitnesssystem/auth/infrastructure/jwt/model/AccessTokenClaims.java
@@ -39,5 +39,5 @@ public class AccessTokenClaims {
     private String username;
 
     /** Token 版本号，用于全局或单用户 Token 失效控制 */
-    private Integer tokenVersion;
+    private Long tokenVersion;
 }
